backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
openai-whisper==20231117
sentence-transformers==2.7.0
torch==2.3.0
transformers==4.40.2
ffmpeg-python==0.2.0
numpy==1.26.4
scipy==1.13.0
yt-dlp
